Move SidePanelBaseManager to internal folder since it's not used outside
