+++ src/core/managers/internal/LedgerConnectStateManager/LedgerConnectStateManager.ts
@@ -9,7 +9,7 @@ import {
   MvxLedgerConnectPanel
 } from 'lib/sdkDappCoreUi';
 import { LedgerConnectEventsEnum } from './types';
-import { SidePanelBaseManager } from '../../SidePanelBaseManager/SidePanelBaseManager';
+import { SidePanelBaseManager } from '../SidePanelBaseManager/SidePanelBaseManager';
 
 export class LedgerConnectStateManager extends SidePanelBaseManager<
   MvxLedgerConnectPanel,
