+++ src/core/managers/internal/PendingTransactionsStateManager/PendingTransactionsStateManager.ts
@@ -4,7 +4,7 @@ import {
   MvxPendingTransactionsPanel
 } from 'lib/sdkDappCoreUi';
 import { PendingTransactionsEventsEnum } from './types/pendingTransactions.types';
-import { SidePanelBaseManager } from '../../SidePanelBaseManager/SidePanelBaseManager';
+import { SidePanelBaseManager } from '../SidePanelBaseManager/SidePanelBaseManager';
 
 export class PendingTransactionsStateManager extends SidePanelBaseManager<
   MvxPendingTransactionsPanel,
