+++ src/core/managers/internal/SidePanelBaseManager/SidePanelBaseManager.ts
