+++ src/core/managers/internal/SidePanelBaseManager/index.ts
